// 04-heroes-app/src/app/heroes/interfaces/hero-interface.ts
export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics"
}

export interface HeroInterface {
    id:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    alt_img?:         string;
}